Fix bug 29116 (quiver "IOError: bad message length" followed by freeze)
There was no coverage against chr1, so it was getting
no-called, which meant a struct was getting sent over the
multiprocessing pickle queue that contained an array of zeros
the length of chr1.  And it was being encoded as a numpy array
of uint64, as opposed to uint8, so it was needlessly consuming
way too much memory, somehow triggering failure in
multiprocessing.

Fix is to encode as uint8 array.

[git-p4: depot-paths = "//depot/software/smrtanalysis/bioinformatics/tools/GenomicConsensus/": change = 165287]

+++ GenomicConsensus/consensus.py
@@ -63,17 +63,17 @@ class Consensus(object):
         length = len(referenceSequence)
         seq = np.empty(length, dtype="S1")
         seq.fill("N")
-        conf = np.zeros(length, dtype=np.uint)
+        conf = np.zeros(length, dtype=np.uint8)
         return cls(refWin, seq.tostring(), conf)
 
     @classmethod
     def referenceAsConsensus(cls, refWin, referenceSequence):
-        conf = np.zeros(len(referenceSequence), dtype=np.uint)
+        conf = np.zeros(len(referenceSequence), dtype=np.uint8)
         return cls(refWin, referenceSequence, conf)
 
     @classmethod
     def lowercaseReferenceAsConsensus(cls, refWin, referenceSequence):
-        conf = np.zeros(len(referenceSequence), dtype=np.uint)
+        conf = np.zeros(len(referenceSequence), dtype=np.uint8)
         return cls(refWin, referenceSequence.lower(), conf)
 
     @classmethod
